+++ src/lc/isValidBST.go
@@ -6,14 +6,14 @@ func IsValidBST(root *TreeNode) bool {
 	return isValidBSTHepler(root, math.MinInt64, math.MaxInt64)
 }
 
-func isValidBSTHepler(root *TreeNode, l,u int) bool {
+func isValidBSTHepler(root *TreeNode, l, u int64) bool {
 	if root == nil {
 		return true
 	}
 
-	if root.Val <= l || root.Val >= u {
+	if int64(root.Val) <= l || int64(root.Val) >= u {
 		return false
 	}
 
-	return isValidBSTHepler(root.Left, l, root.Val) && isValidBSTHepler(root.Right, root.Val, u)
+	return isValidBSTHepler(root.Left, l, int64(root.Val)) && isValidBSTHepler(root.Right, int64(root.Val), u)
 }
